인풋개선(중량 소수점 입력 마이그레이션)

+++ app/src/main/java/com/progressive/overload/model/entity/RecordTable.kt
@@ -22,7 +22,7 @@ data class RecordTable(
     @ColumnInfo
     val name: String?,
     @ColumnInfo
-    val weight: Int?,
+    val weight: Float?,
     @ColumnInfo
     val repitition: Int?,
     @ColumnInfo
